@@ -32,19 +32,19 @@
     "@lingui/react": "^3.8.9",
     "@monaco-editor/react": "^4.4.6",
     "@notionhq/client": "^0.4.13",
-    "@radix-ui/react-alert-dialog": "^1.0.2",
-    "@radix-ui/react-collapsible": "^1.0.2",
-    "@radix-ui/react-dialog": "^1.0.3",
-    "@radix-ui/react-dropdown-menu": "^2.0.1",
-    "@radix-ui/react-hover-card": "^0.1.5",
-    "@radix-ui/react-navigation-menu": "^1.0.0",
-    "@radix-ui/react-popover": "^1.0.3",
-    "@radix-ui/react-progress": "^1.0.2",
-    "@radix-ui/react-radio-group": "^1.1.2",
-    "@radix-ui/react-select": "^1.2.1",
-    "@radix-ui/react-slider": "^1.1.0",
-    "@radix-ui/react-tabs": "^1.0.1",
-    "@radix-ui/react-toggle": "^0.1.4",
+    "@radix-ui/react-alert-dialog": "^1.0.4",
+    "@radix-ui/react-collapsible": "^1.0.3",
+    "@radix-ui/react-dialog": "^1.0.4",
+    "@radix-ui/react-dropdown-menu": "^2.0.5",
+    "@radix-ui/react-hover-card": "^1.0.6",
+    "@radix-ui/react-navigation-menu": "^1.1.3",
+    "@radix-ui/react-popover": "^1.0.6",
+    "@radix-ui/react-progress": "^1.0.3",
+    "@radix-ui/react-radio-group": "^1.1.3",
+    "@radix-ui/react-select": "^1.2.2",
+    "@radix-ui/react-slider": "^1.1.2",
+    "@radix-ui/react-tabs": "^1.0.4",
+    "@radix-ui/react-toggle": "^1.0.3",
     "@radix-ui/react-tooltip": "^1.0.6",
     "@react-hook/throttle": "^2.2.0",
     "@sendgrid/mail": "^7.4.6",
@@ -62,8 +62,6 @@
     "@testing-library/user-event": "^12.1.10",
     "@tone-row/slang": "^1.2.35",
     "@tone-row/strip-comments": "^2.0.1",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
     "buffer": "^6.0.3",
     "classnames": "^2.3.2",
     "core-js": "^3.18.1",
@@ -99,17 +97,17 @@
     "postcss-flexbugs-fixes": "^5.0.2",
     "prettier": "^2.3.1",
     "re-resizable": "^6.9.0",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-confetti": "^6.1.0",
     "react-contexify": "^5.0.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-ga": "^3.3.0",
     "react-gtm-module": "^2.0.11",
     "react-hook-form": "^7.0.7",
     "react-icons": "^4.9.0",
     "react-query": "^3.26.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.14.2",
     "react-scripts": "5.0.1",
     "react-use-localstorage": "^3.5.3",
     "shared": "workspace:*",
@@ -158,6 +156,8 @@
     "@types/node-localstorage": "^1.3.0",
     "@types/pako": "^2.0.0",
     "@types/papaparse": "^5.3.7",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
     "@types/react-gtm-module": "^2.0.1",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-select": "^4.0.17",
