@@ -2,7 +2,7 @@ import { t, Trans } from "@lingui/macro";
 import produce from "immer";
 import { Palette } from "phosphor-react";
 import { FaRegSnowflake } from "react-icons/fa";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import { GraphOptionsObject } from "../../lib/constants";
 import { defaultLayout, getLayout } from "../../lib/getLayout";
@@ -63,7 +63,7 @@ export function EditLayoutTab() {
     spacingFactor = layout.spacingFactor;
   }
 
-  const { push } = useHistory();
+  const navigate = useNavigate();
 
   if (isFrozen) return <FrozenLayout />;
 
@@ -194,7 +194,7 @@ export function EditLayoutTab() {
           size="md"
           rightIcon={<Palette size={20} />}
           className="ml-5 mr-1"
-          onClick={() => push("/pricing")}
+          onClick={() => navigate("/pricing")}
         >
           <Trans>Get More Layouts</Trans>
         </Button2>
