@@ -5,8 +5,7 @@ import "core-js/features/object/from-entries";
 import "core-js/features/object/entries";
 
 import { Buffer } from "buffer";
-import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import App from "./components/App";
 import { initSentry } from "./lib/sentry";
@@ -22,12 +21,11 @@ window.Buffer = Buffer;
 
 initSentry();
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById("root")
-);
+const container = document.getElementById("root");
+if (!container) throw new Error("No root element");
+const root = createRoot(container);
+
+root.render(<App />);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
