@@ -1,6 +1,6 @@
 import { decompressFromEncodedURIComponent } from "lz-string";
 import { useQuery } from "react-query";
-import { useParams, useRouteMatch } from "react-router-dom";
+import { useLocation, useParams } from "react-router-dom";
 
 import EditorError from "../components/EditorError";
 import { EditorWrapper } from "../components/EditorWrapper";
@@ -11,15 +11,19 @@ import { prepareChart } from "../lib/prepareChart/prepareChart";
 import { useDoc } from "../lib/useDoc";
 
 function ReadOnly() {
-  const { path } = useRouteMatch();
+  const { pathname } = useLocation();
   const { graphText = window.location.hash.slice(1) } = useParams<{
     graphText: string;
   }>();
-  useQuery(["read", path, graphText], () => loadReadOnly(path, graphText), {
-    enabled: typeof graphText === "string",
-    suspense: true,
-    staleTime: 0,
-  });
+  useQuery(
+    ["read", pathname, graphText],
+    () => loadReadOnly(pathname, graphText),
+    {
+      enabled: typeof graphText === "string",
+      suspense: true,
+      staleTime: 0,
+    }
+  );
 
   const text = useDoc((d) => d.text);
 
@@ -43,12 +47,7 @@ function ReadOnly() {
 export default ReadOnly;
 
 async function loadReadOnly(path: string, graphText: string) {
-  const isCompressed = [
-    "/c/:graphText?",
-    "/f/:graphText?",
-    "/n/:graphText?",
-  ].includes(path);
-  const initialText = isCompressed
+  const initialText = isCompressed(path)
     ? decompressFromEncodedURIComponent(graphText) ?? ""
     : decodeURIComponent(graphText);
   await prepareChart(initialText, {
@@ -57,4 +56,8 @@ async function loadReadOnly(path: string, graphText: string) {
     title: "read-only-flowchart",
   });
   return initialText;
+}
+
+function isCompressed(pathname: string) {
+  return ["/c/", "/f/", "/n/"].some((p) => pathname.startsWith(p));
 }
